[checkpointing][refactor] Extract ChannelStatePersister
Motivation: unit testing
